+++ odata_query/sql/__init__.py
@@ -1,2 +1,3 @@
 from .athena import AstToAthenaSqlVisitor
+from .base import AstToSqlVisitor
 from .sqlite import AstToSqliteSqlVisitor