+++ CHANGELOG.rst
@@ -8,6 +8,24 @@ The format is based on `Keep a Changelog <https://keepachangelog.com/en/1.0.0/>`
 and this project adheres to `Semantic Versioning <https://semver.org/spec/v2.0.0.html>`_.
 
 
+[Unreleased]
+------------
+
+Added
+^^^^^
+
+* SQL: Support for the SQLite dialect.
+
+Changed
+^^^^^^^
+
+* SQL: Refactored SQL support
+
+    - The ``AstToSqlVisitor`` now implements basic, standard SQL and functions
+      as a base class for dialects.
+    - The Athena dialect now lives in ``AstToAthenaSqlVisitor``.
+
+
 [0.7.2] - 2022-12-19
 --------------------